Drone's on Parcel
This code Displays "qxg_drone.glb" as a drone. There are 2 things happening. 

The first script makes a drone fly in a circle at an elevated level.

the second script spawns 5 drones. They all are stagnant until someone runs up to it. The drone that is touched then rotates and moves a various distance on parcel.